do not stream logs for now
